requirements.txt
Beaker==1.6.4
Chameleon==2.18
FormEncode==1.3.0a1
Mako==0.9.1
MarkupSafe==0.23
PasteDeploy==1.5.2
Pillow==2.6.1
PyMySQL==0.6.1
PyPages==0.1
Pygments==1.6
SQLAlchemy==0.9.8
WebOb==1.4
argparse==1.2.2
bbcode==1.0.19
mysql-connector-python==2.0.2
numpy==1.9.1
ply==3.4
pymongo==2.7.2
pyramid==1.5.2
pyramid-beaker==0.8
pyramid-chameleon==0.3
pyramid-debugtoolbar==2.0.2
pyramid-mailer==0.13
pyramid-mako==1.0.1
pyramid-redis-sessions==1.0a1
pyramid-tm==0.7
pyramid-whoauth==0.1.2
redis==2.9.1
repoze.lru==0.6
repoze.sendmail==4.2
repoze.who==2.2
six==1.8.0
slimit==0.8.1
## !! Could not determine repository location
testproject==0.0
transaction==1.4.3
translationstring==1.3
unittest2==0.8.0
venusian==1.0
waitress==0.8.8
zope.deprecation==4.1.1
zope.interface==4.1.1
zope.sqlalchemy==0.7.5
